Fix lambda in tgram_docs_bot.py (#56)
the is error:
^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
TypeError: <lambda>() takes 1 positional argument but 2 were given
^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
Inside filters.Filter(lambda m: ...), tgram passes bot and message to the function, but lambda m: ... only expects m, which causes the error. By modifying it to lambda _, m: ..., we show that lambda receives bot as the first parameter and m as the second parameter, and thus it runs without problems.